# [WIP] Faster Expectation Calculation