--- a/pymbar/mbar.py
+++ b/pymbar/mbar.py
@@ -620,13 +620,18 @@ class MBAR:
         N_k[0:K] = self.N_k
         f_k[0:K] = self.f_k
 
+        # Pre-calculate some quantities outside the for loops
+        states_with_samples = (self.N_k > 0)
+        log_denominator_n = logsumexp(self.f_k[states_with_samples] - self.u_kn[states_with_samples].T, b=self.N_k[states_with_samples], axis=1)
         # Compute row of W_nk matrix for the extra states corresponding to u_ln 
         # that the state list specifies
         for l in L_list:
             la = K+l  #l, augmented
-            Log_W_nk[:, la] = self._computeUnnormalizedLogWeights(u_ln[l,:]) 
-            f_k[la] = -logsumexp(Log_W_nk[:, la])
-            Log_W_nk[:, la] += f_k[la]
+            # Calculate log normalizing constants and log weights via Eqns 13, 14
+            log_C_a = -logsumexp(-u_ln[l] - log_denominator_n)
+            Q = log_C_a - u_ln[l] - log_denominator_n
+            Log_W_nk[:, la] = Q
+            f_k[la] = log_C_a
 
         # Compute the remaining rows/columns of W_nk, and calculate
         # their normalizing constants c_k