Added expectation performance fix and timing code, as well as helper functions for testsystem generation.

For the master branch, we see this:

```
In [23]: %time A, dA = mbar.computeExpectations(x, compute_uncertainty=False)
CPU times: user 29.8 s, sys: 4.65 s, total: 34.5 s
Wall time: 9.65 s

In [24]: %time A = mbar0.computeExpectations(x, compute_uncertainty=False)
CPU times: user 111 ms, sys: 3.4 ms, total: 115 ms
Wall time: 114 ms
```

The current pull request reduces this to:

```
In [23]: %time A, dA = mbar.computeExpectations(x, compute_uncertainty=False)
CPU times: user 515 ms, sys: 70.8 ms, total: 586 ms
Wall time: 342 ms

In [24]: %time A = mbar0.computeExpectations(x, compute_uncertainty=False)
CPU times: user 111 ms, sys: 3.72 ms, total: 115 ms
Wall time: 114 ms
```