@kyleabeauchamp 2015-02-18
OK, made those two fixes