@mrshirts 2015-02-18
This is great!  Thanks for recognizing the inefficiency.  The good thing about a single inner loop setup is that it only needs to be fixed in one place.