@kyleabeauchamp 2015-02-18
We're still not quite matching the performance of MBAR 2.0, but it's within a factor of 3, which is good enough for expectations, which aren't rate limiting in most cases anyway.