@mrshirts 2015-02-18
Sounds good - we can continue to make tweaks as we go.  Indeed,
expectations should not be time limiting.  I can look over the timings and
help tweak as the paper is prepared.

On Wed, Feb 18, 2015 at 12:35 PM, Kyle Beauchamp <notifications@github.com>
wrote:

> We're still not quite matching the performance of MBAR 2.0, but it's
> within a factor of 3, which is good enough for expectations, which aren't
> rate limiting in most cases anyway.
>
> —
> Reply to this email directly or view it on GitHub
> <https://github.com/choderalab/pymbar/pull/171#issuecomment-74908708>.
>